+++ packages/opencode-extras/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamba/opencode-extras
 
+## 1.4.0
+
+### Patch Changes
+
+- @chamba/adapters@1.4.0
+- @chamba/core@1.4.0
+
 ## 1.3.0
 
 ### Minor Changes